# Minor speedup