@etiennebacher 2024-01-12
Thanks for the offer, but I just don't think this tiny addition is enough for me to be listed as contributor (nothing against your package, it's great).

Feel free to merge :)